requirements.txt
# Web Framework & Auth
Flask==3.1.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.1.4

# AI & LLM
groq==0.37.1
python-dotenv==1.2.1

# Document & Image Processing
easyocr==1.7.2
PyMuPDF==1.26.7
pypdf==6.4.2
pillow==12.0.0
opencv-python-headless==4.12.0.88

# PDF Generation & Data
reportlab==4.4.7
pandas==2.3.3
numpy==2.2.6

# Database & Utilities
SQLAlchemy==2.0.45
requests==2.32.5
gunicorn==23.0.0